+++ dart/geobase/lib/src/vector_data/coords/lonlat.dart
@@ -23,7 +23,7 @@ import '/src/vector_data/array.dart';
 @immutable
 class LonLat extends PositionCoords implements Geographic {
   /// A geographic position as an iterable collection of [lon] and [lat] values.
-  /// 
+  ///
   /// Longitude is normalized using [normalizeLongitude] and latitude is clamped
   /// using [clampLatitude] before storing values.
   factory LonLat(double lon, double lat) {
@@ -158,7 +158,7 @@ class LonLat extends PositionCoords implements Geographic {
 class LonLatElev extends LonLat {
   /// A geographic position as an iterable collection of [lon], [lat] and [elev]
   /// values.
-  /// 
+  ///
   /// Longitude is normalized using [normalizeLongitude] and latitude is clamped
   /// using [clampLatitude] before storing values.
   factory LonLatElev(double lon, double lat, double elev) {
@@ -260,7 +260,7 @@ class LonLatElev extends LonLat {
 class LonLatM extends LonLat {
   /// A geographic position as an iterable collection of [lon], [lat] and [m]
   /// values.
-  /// 
+  ///
   /// Longitude is normalized using [normalizeLongitude] and latitude is clamped
   /// using [clampLatitude] before storing values.
   factory LonLatM(double lon, double lat, double m) {
@@ -356,7 +356,7 @@ class LonLatM extends LonLat {
 class LonLatElevM extends LonLatElev {
   /// A geographic position as an iterable collection of [lon], [lat], [elev]
   /// and [m] values.
-  /// 
+  ///
   /// Longitude is normalized using [normalizeLongitude] and latitude is clamped
   /// using [clampLatitude] before storing values.
   factory LonLatElevM(double lon, double lat, double elev, double m) {